requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database Clients
psycopg2-binary==2.9.9
pymongo==4.6.1
pymilvus==2.3.5

# Kafka & Streaming
kafka-python==2.0.2
confluent-kafka==2.3.0

# AI & ML
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5
langgraph==0.0.20
tiktoken==0.5.2

# OCR
requests==2.31.0
Pillow==10.2.0
PyPDF2==3.0.1

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
aiofiles==23.2.1
httpx==0.26.0
PyYAML==6.0.1

# Caching
redis==5.0.1
hiredis==2.3.2

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
numpy==1.26.3

# Type Checking
mypy==1.8.0
types-requests==2.31.0.20240125
